installer: Avoid overwriting files if they are already present.

+++ k8s-installer/installer.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
-cp -f /bin/sriov /host-cni-bin/
-cp -f /bin/fixipam /host-cni-bin/
+if [ ! -e /host-cni-bin/sriov ]; then
+	cp /bin/sriov /host-cni-bin/
+fi
 
-cp -f /installer/10-sriov-cni.conf /host-cni-etc/
+if [ ! -e /host-cni-bin/fixipam ]; then
+	cp -f /bin/fixipam /host-cni-bin/
+fi
+
+if [ ! -e /host-cni-etc/10-sriov-cni.conf ]; then
+	cp /installer/10-sriov-cni.conf /host-cni-etc/
+fi
